% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISBDZDL.R
\docType{package}
\name{ISBDZDL}
\alias{ISBDZDL}
\title{ISBDZDL: An R Package Demo.}
\description{
This is a package for generating mixture-Gaussian or mixture-t
random variables/random vectors.
}
\section{Mixed-distribution random variables generation functions}{

The mixed-distribution random variables generation functions can generate
mixed-Gaussian or mixed-t random variables/vectors.
}

