% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_mixed.R
\name{r_mixed}
\alias{r_mixed}
\alias{r_gaussian_mixed}
\alias{r_t_mixed}
\title{Generate mixed-distribution random variables}
\usage{
r_gaussian_mixed(n = 100, mu, sigma, weight, tag = FALSE)

r_t_mixed(n = 100, sigma, df, mu, weight, tag = FALSE)
}
\arguments{
\item{n}{number of random variables/vectors to be generated.}

\item{mu}{matrix with each column the mean vector of each Gaussian component.}

\item{sigma}{array consisting of the sigma matrices of Gaussian components.}

\item{weight}{vector of the weights for Gaussian components.}

\item{tag}{logical; if FALSE(default), the matrix returned will not contain
an extra column of class label.}

\item{df}{degree of freedom for t-distribution}
}
\value{
A matrix of mixed-Gaussian(or mixed-t) random variables/vectors
with or without class labels.
}
\description{
\code{r_gaussian_mixed}(or \code{r_t_mixed}) returns a matrix of
mixed-Gaussian(or mixed-t) random variables/vectors
with or without class labels.
}
\details{
The default value for \code{n} is 100.

If \code{tag} is \code{TRUE}, there will be an extra column added at the
end of the output matrix representing the component to which each
random variable or vector belongs.
}
\examples{
mu <- matrix(rep(c(-1, 0, 1), 4), byrow = T, nrow = 4)
sigma <- matrix(rep(0, 4 * 4), nrow = 4)
for (i in 1:(4 - 1)) {
  for (j in (i + 1):4) {
    sigma[i, j] <- 0.5^ (j - i)
  }
}
sigma <- sigma + t(sigma) + diag(1, nrow = 4)
sigma <- array(rep(sigma, 3), dim = c(4, 4, 3))
weight <- rep(1, 3) / 3
r_g <- r_gaussian_mixed(n = 100, mu, sigma, weight, tag = FALSE)
r_t <- r_t_mixed(n = 100, sigma, df = 1, mu, weight, tag = FALSE)
}
